Move constants to platform
