+++ src/kernels/ipywidgets-message-coordination/ipyWidgetMessageDispatcher.ts
@@ -8,13 +8,13 @@ import * as uuid from 'uuid/v4';
 import { Event, EventEmitter, NotebookDocument } from 'vscode';
 import type { Data as WebSocketData } from 'ws';
 import { traceVerbose, traceError, traceInfo } from '../../platform/logging';
+import { Identifiers } from '../../platform/common/constants';
 import { IDisposable } from '../../platform/common/types';
 import { Deferred, createDeferred } from '../../platform/common/utils/async';
 import { noop } from '../../platform/common/utils/misc';
 import { deserializeDataViews, serializeDataViews } from '../../platform/common/utils/serializers';
 import { IPyWidgetMessages, IInteractiveWindowMapping } from '../../platform/messageTypes';
 import { sendTelemetryEvent, Telemetry } from '../../telemetry';
-import { Identifiers } from '../../webviews/webview-side/common/constants';
 import { IKernel, IKernelProvider, KernelSocketInformation } from '../types';
 import { WIDGET_MIMETYPE } from './constants';
 import { IIPyWidgetMessageDispatcher, IPyWidgetMessage } from './types';
