+++ src/kernels/jupyter/import-export/jupyterExporter.node.ts
@@ -17,10 +17,11 @@ import { ICell, IConfigurationService } from '../../../platform/common/types';
 import { concatMultilineString, pruneCell } from '../../../platform/common/utils';
 import { DataScience } from '../../../platform/common/utils/localize';
 import { IDataScienceErrorHandler } from '../../../platform/errors/types';
-import { defaultNotebookFormat, CodeSnippets } from '../../../webviews/webview-side/common/constants';
+import { defaultNotebookFormat } from '../../../platform/common/constants';
 import { INotebookExporter, IJupyterExecution } from '../types';
 import { openAndShowNotebook } from '../../../platform/common/utils/notebooks';
 import { noop } from '../../../platform/common/utils/misc';
+import { CodeSnippets } from '../../../platform/common/constants';
 
 @injectable()
 export class JupyterExporter implements INotebookExporter {
