+++ src/kernels/jupyter/import-export/jupyterImporter.node.ts
@@ -10,10 +10,10 @@ import * as path from '../../../platform/vscode-path/path';
 import { Uri } from 'vscode';
 import { IWorkspaceService } from '../../../platform/common/application/types';
 import { traceError } from '../../../platform/logging';
+import { Identifiers, CodeSnippets } from '../../../platform/common/constants';
 import { IPlatformService } from '../../../platform/common/platform/types';
 import { IDisposableRegistry, IConfigurationService } from '../../../platform/common/types';
 import { DataScience } from '../../../platform/common/utils/localize';
-import { Identifiers, CodeSnippets } from '../../../webviews/webview-side/common/constants';
 import { PythonEnvironment } from '../../../platform/pythonEnvironments/info';
 import { noop } from '../../../platform/common/utils/misc';
 import { INotebookImporter, INbConvertInterpreterDependencyChecker, INbConvertExportToPythonService } from '../types';
