+++ src/kernels/jupyter/interpreter/jupyterCommand.node.ts
@@ -13,7 +13,7 @@ import {
 } from '../../../platform/common/process/types.node';
 import { EXTENSION_ROOT_DIR } from '../../../platform/constants.node';
 import { PythonEnvironment } from '../../../platform/pythonEnvironments/info';
-import { JupyterCommands, JupyterDaemonModule } from '../../../webviews/webview-side/common/constants';
+import { JupyterCommands, JupyterDaemonModule } from '../../../platform/common/constants';
 import { IJupyterCommand, IJupyterCommandFactory } from '../types.node';
 
 class InterpreterJupyterCommand implements IJupyterCommand {
