+++ src/kernels/jupyter/session/jupyterSessionManagerFactory.ts
@@ -4,8 +4,8 @@
 import { inject, injectable, named, optional } from 'inversify';
 import { JupyterSessionManager } from './jupyterSessionManager';
 import { IApplicationShell } from '../../../platform/common/application/types';
+import { JUPYTER_OUTPUT_CHANNEL } from '../../../platform/common/constants';
 import { IConfigurationService, IOutputChannel, IPersistentStateFactory } from '../../../platform/common/types';
-import { JUPYTER_OUTPUT_CHANNEL } from '../../../webviews/webview-side/common/constants';
 import { IJupyterConnection } from '../../types';
 import {
     IJupyterSessionManagerFactory,
