+++ src/kernels/variables/jupyterVariables.ts
@@ -5,8 +5,8 @@ import { inject, injectable, named } from 'inversify';
 
 import { CancellationToken, Event, EventEmitter } from 'vscode';
 import { IDisposableRegistry } from '../../platform/common/types';
+import { Identifiers } from '../../platform/common/constants';
 import { captureTelemetry, Telemetry } from '../../telemetry';
-import { Identifiers } from '../../webviews/webview-side/common/constants';
 import { IKernel } from '../types';
 import {
     IJupyterVariables,
