+++ src/platform/common/constants.ts
@@ -32,6 +32,9 @@ export const STANDARD_OUTPUT_CHANNEL = 'STANDARD_OUTPUT_CHANNEL';
 
 export * from '../constants';
 
+/**
+ * TODO@rebornix, move webviews/webview-side/common/constants to right places
+ */
 export * from '../../webviews/webview-side/common/constants';
 
 export namespace HelpLinks {
@@ -79,3 +82,47 @@ export function isUnitTestExecution(): boolean {
 export function setUnitTestExecution(enabled: boolean) {
     _isUnitTestExecution = enabled;
 }
+
+export namespace Identifiers {
+    export const GeneratedThemeName = 'ipython-theme'; // This needs to be all lower class and a valid class name.
+    export const RawPurpose = 'raw';
+    export const MatplotLibDefaultParams = '_VSCode_defaultMatplotlib_Params';
+    export const MatplotLibFigureFormats = '_VSCode_matplotLib_FigureFormats';
+    export const DefaultCodeCellMarker = '# %%';
+    export const DefaultCommTarget = 'jupyter.widget';
+    export const ALL_VARIABLES = 'ALL_VARIABLES';
+    export const KERNEL_VARIABLES = 'KERNEL_VARIABLES';
+    export const DEBUGGER_VARIABLES = 'DEBUGGER_VARIABLES';
+    export const PYTHON_VARIABLES_REQUESTER = 'PYTHON_VARIABLES_REQUESTER';
+    export const MULTIPLEXING_DEBUGSERVICE = 'MULTIPLEXING_DEBUGSERVICE';
+    export const RUN_BY_LINE_DEBUGSERVICE = 'RUN_BY_LINE_DEBUGSERVICE';
+    export const REMOTE_URI = 'https://remote/';
+    export const REMOTE_URI_ID_PARAM = 'id';
+    export const REMOTE_URI_HANDLE_PARAM = 'uriHandle';
+}
+
+export namespace CodeSnippets {
+    export const ChangeDirectory = [
+        '{0}',
+        '{1}',
+        'import os',
+        'try:',
+        "\tos.chdir(os.path.join(os.getcwd(), '{2}'))",
+        '\tprint(os.getcwd())',
+        'except:',
+        '\tpass',
+        ''
+    ];
+    export const ChangeDirectoryCommentIdentifier = '# ms-toolsai.jupyter added'; // Not translated so can compare.
+    export const ImportIPython = '{0}\nfrom IPython import get_ipython\n\n{1}';
+    export const MatplotLibInit = `import matplotlib\n%matplotlib inline\n${Identifiers.MatplotLibDefaultParams} = dict(matplotlib.rcParams)\n`;
+    export const AppendSVGFigureFormat = `import matplotlib_inline.backend_inline\n${Identifiers.MatplotLibFigureFormats} = matplotlib_inline.backend_inline.InlineBackend.instance().figure_formats\n${Identifiers.MatplotLibFigureFormats}.add('svg')\nmatplotlib_inline.backend_inline.set_matplotlib_formats(*${Identifiers.MatplotLibFigureFormats})`;
+    export const UpdateCWDAndPath =
+        'import os\nimport sys\n%cd "{0}"\nif os.getcwd() not in sys.path:\n    sys.path.insert(0, os.getcwd())';
+    export const DisableJedi = '%config Completer.use_jedi = False';
+}
+
+// Identifier for the output panel that will display the output from the Jupyter Server.
+export const JUPYTER_OUTPUT_CHANNEL = 'JUPYTER_OUTPUT_CHANNEL';
+export const KernelInterruptDaemonModule = 'vscode_datascience_helpers.kernel_interrupt_daemon';
+export const JupyterDaemonModule = 'vscode_datascience_helpers.jupyter_daemon';
